Combined IncomingResponse, OutgoingResponse, and ResponseTrait
Created Response from the three previous classes, using the
MessageTrait, and encapsulating all concerns of a Response.
