+++ src/Response.php
@@ -1,11 +1,17 @@
 <?php
 namespace Phly\Http;
 
+use InvalidArgumentException;
+use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\StreamableInterface;
+
 /**
- * Common (accessor) methods for all HTTP responses.
+ * HTTP response encapsulation.
  */
-trait ResponseTrait
+class Response implements ResponseInterface
 {
+    use MessageTrait;
+
     /**
      * Map of standard HTTP status code/reason phrases
      *
@@ -87,6 +93,26 @@ trait ResponseTrait
     private $statusCode = 200;
 
     /**
+     * @param string|resource|StreamableInterface $stream Stream identifier and/or actual stream resource
+     */
+    public function __construct($stream = 'php://memory')
+    {
+        if (! is_string($stream) && ! is_resource($stream) && ! $stream instanceof StreamableInterface) {
+            throw new InvalidArgumentException(
+                'Stream must be a string stream resource identifier, '
+                . 'an actual stream resource, '
+                . 'or a Psr\Http\Message\StreamableInterface implementation'
+            );
+        }
+
+        if (! $stream instanceof StreamableInterface) {
+            $stream = new Stream($stream, 'wb+');
+        }
+
+        $this->setBody($stream);
+    }
+
+    /**
      * Gets the response Status-Code.
      *
      * The Status-Code is a 3-digit integer result code of the server's attempt
@@ -118,5 +144,31 @@ trait ResponseTrait
         }
 
         return $this->reasonPhrase;
+    }
+
+    /**
+     * Sets the status code of this response.
+     *
+     * @param integer $code The 3-digit integer result code to set.
+     * @param null|string $reasonPhrase The reason phrase to use with the status provided;
+     *     if none is provided, and the status has a match in ResponseTrait::$phrases, the
+     *     corresponding value will be used.
+     * @return void
+     */
+    public function setStatus($code, $reasonPhrase = null)
+    {
+        if (! is_numeric($code)
+            || is_float($code)
+            || $code < 100
+            || $code >= 600
+        ) {
+            throw new InvalidArgumentException(sprintf(
+                'Invalid status code "%s"; must be an integer between 100 and 599, inclusive',
+                (is_scalar($code) ? $code : gettype($code))
+            ));
+        }
+
+        $this->statusCode   = (int) $code;
+        $this->reasonPhrase = $reasonPhrase;
     }
 }
